move usage.go to pkg/cli for better reuse
