+++ pkg/cli/usage.go
@@ -1,4 +1,4 @@
-package command
+package cli
 
 import (
 	"fmt"
